releasetools/*image.sh
 - Removed an "export CPPFLAGS=${FLAG}" which prevented proper settings
   from the environment to be passed to the build step.

 - Re-arranged variable settings at the top, so that we don't need
   to evaluate the argument to vfat mkfs command anymore in
   arm_sdimage.sh

 - "Merged" both script for easier comparison in the future.

+++ releasetools/arm_sdimage.sh
@@ -2,6 +2,11 @@
 set -e
 
 #
+# This script creates a bootable image and should at some point in the future
+# be replaced by makefs.
+#
+
+#
 # Source settings if present
 #
 : ${SETTINGS_MINIX=.settings}
@@ -24,32 +29,53 @@ fi
 : ${FSTAB=${DESTDIR}/etc/fstab}
 : ${BUILDVARS=-V MKGCCCMDS=yes -V MKLLVM=no}
 : ${BUILDSH=build.sh}
+: ${CREATE_IMAGE_ONLY=0}
+: ${RC=minix_x86.rc}
+
+# Where the kernel & boot modules will be
+MODDIR=${DESTDIR}/boot/minix/.temp
 
 #
 # Directory where to store temporary file system images
 #
 : ${IMG_DIR=${OBJ}/img}
-: ${IMG=minix_arm_sd.img}
 : ${MLO=MLO}
 : ${UBOOT=u-boot.img}
-
 
 # Beagleboard-xm
 : ${U_BOOT_BIN_DIR=build/omap3_beagle/}
 : ${CONSOLE=tty02}
-
 
 # BeagleBone (and black)
 #: ${U_BOOT_BIN_DIR=build/am335x_evm/}
 #: ${CONSOLE=tty00}
 
 #
-#
 # We host u-boot binaries.
+#
 U_BOOT_GIT_VERSION=cb5178f12787c690cb1c888d88733137e5a47b15
 
-# Set the vfat mkfs command. We need to re-evaluate MKFS_VFAT_OPTS after
-# FAT_SIZE is set :-(
+#
+# All sized are written in 512 byte blocks
+#
+# we create a disk image of about 2 gig's
+# for alignment reasons, prefer sizes which are multiples of 4096 bytes
+#
+: ${IMG_SIZE=$((     2*(2**30) / 512))}
+: ${FAT_SIZE=$((    10*(2**20) / 512))}
+: ${ROOT_SIZE=$((   64*(2**20) / 512))}
+: ${HOME_SIZE=$((  128*(2**20) / 512))}
+: ${USR_SIZE=$((  1792*(2**20) / 512))}
+
+#
+# Do some math to determine the start addresses of the partitions.
+# Don't leave holes so the 'partition' invocation later is easy.
+#
+FAT_START=2048
+ROOT_START=$(($FAT_START + $FAT_SIZE))
+USR_START=$(($ROOT_START + $ROOT_SIZE))
+HOME_START=$(($USR_START + $USR_SIZE))
+
 case $(uname -s) in
 Darwin)
 	MKFS_VFAT_CMD=newfs_msdos
@@ -65,14 +91,13 @@ FreeBSD)
 ;;
 esac
 
-
 if [ -n "$BASE_URL" ]
 then
 	#we no longer download u-boot but do a checkout
 	#BASE_URL used to be the base url for u-boot
 	#Downloads
 	echo "Warning:** Setting BASE_URL (u-boot) is no longer possible use U_BOOT_BIN_DIR"
-	echo "Look in ./releasetools/arm_sdimage.sh for suggested values"
+	echo "Look in ${RELEASETOOLSDIR}/arm_sdimage.sh for suggested values"
 	exit 1
 fi
 
@@ -82,7 +107,7 @@ then
 	exit 1
 fi
 
-export PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:$PATH
+export PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:${PATH}
 
 for needed in mcopy dd ${MKFS_VFAT_CMD} git
 do
@@ -93,52 +118,58 @@ do
 	fi
 done
 
+: ${IMG=minix_arm_sd.img}
+
+#
+# Are we going to build the minix sources?
+#
+
+if [ ${CREATE_IMAGE_ONLY} -eq 1 ]
+then
+	if [ ! -d ${DESTDIR} ]
+	then
+		echo "Minix source code does'nt appear to have been built."
+		echo "Please try with \$CREATE_IMAGE_ONLY set to 0."
+		exit 1
+	fi
+fi
+
 #
 # Artifacts from this script are stored in the IMG_DIR
 #
-mkdir -p $IMG_DIR
+rm -rf ${IMG_DIR} ${IMG}
+mkdir -p ${IMG_DIR}
 
 #
 # Download the stage 1 bootloader  and u-boot
 #
-./releasetools/fetch_u-boot.sh -o ${RELEASETOOLSDIR}/u-boot -n $U_BOOT_GIT_VERSION
+${RELEASETOOLSDIR}/fetch_u-boot.sh -o ${RELEASETOOLSDIR}/u-boot -n $U_BOOT_GIT_VERSION
 cp ${RELEASETOOLSDIR}/u-boot/${U_BOOT_BIN_DIR}/u-boot.img ${IMG_DIR}/
 cp ${RELEASETOOLSDIR}/u-boot/${U_BOOT_BIN_DIR}/MLO ${IMG_DIR}/
 
-#
-# Remove the generated files to allow us call build.sh without '-V SLOPPY_FLIST=yes'.
-#
-rm -f ${FSTAB}
+if [ ${CREATE_IMAGE_ONLY} -eq 0 ]
+then
+	echo "Going to build Minix source code..."
+	#
+	# Remove the generated files to allow us call build.sh without '-V SLOPPY_FLIST=yes'.
+	#
+	rm -f ${FSTAB}
 
-#
-# Now start the build
-#
-export CPPFLAGS=${FLAG}
-sh ${BUILDSH} -j ${JOBS} -m ${ARCH} -O ${OBJ} -D ${DESTDIR} ${BUILDVARS} -U -u distribution
+	#
+	# Now start the build.
+	#
+	sh ${BUILDSH} -j ${JOBS} -m ${ARCH} -O ${OBJ} -D ${DESTDIR} ${BUILDVARS} -U -u distribution
 
-#
-# This script creates a bootable image and should at some point in the future
-# be replaced by makefs.
-#
-# All sized are written in 512 byte blocks
-#
-# we create a disk image of about 2 gig's
-: ${IMG_SIZE=$((     2*(2**30) / 512))}
-: ${FAT_SIZE=$((    10*(2**20) / 512))}
-: ${ROOT_SIZE=$((   64*(2**20) / 512))}
-: ${HOME_SIZE=$((  128*(2**20) / 512))}
-: ${USR_SIZE=$((  1536*(2**20) / 512))}
-
-eval MKFS_VFAT_OPTS="\"$MKFS_VFAT_OPTS\""
+fi
 
 #
 # create a fstab entry in /etc this is normally done during the
 # setup phase on x86
 #
 cat >${FSTAB} <<END_FSTAB
-/dev/c0d0p2   /home   mfs     rw                      0       2
-/dev/c0d0p3   /usr    mfs     rw                      0       2
-none          /sys    devman  rw,rslabel=devman       0       0
+/dev/c0d0p2	/usr	mfs	rw			0	2
+/dev/c0d0p3	/home	mfs	rw			0	2
+none		/sys	devman	rw,rslabel=devman	0	0
 END_FSTAB
 
 rm -f ${DESTDIR}/SETS.*
@@ -146,31 +177,42 @@ rm -f ${DESTDIR}/SETS.*
 ${CROSS_TOOLS}/nbpwd_mkdb -V 0 -p -d ${DESTDIR} ${DESTDIR}/etc/master.passwd
 
 #
-# Now given the sizes above use DD to create separate files representing
-# the partitions we are going to use.
+# make the different file system. this part is *also* hacky. We first convert
+# the METALOG.sanitised using mtree into a input METALOG containing uids and
+# gids.
+# After that we do some magic processing to add device nodes (also missing from METALOG)
+# and convert the METALOG into a proto file that can be used by mkfs.mfs
+#
+echo "Creating the file systems"
+
+#
+# read METALOG and use mtree to convert the user and group names into uid and gids
+# FIX put "input somewhere clean"
+#
+cat ${DESTDIR}/METALOG.sanitised | ${CROSS_TOOLS}/nbmtree -N ${DESTDIR}/etc -C -K device > ${IMG_DIR}/input
+
+# add rc (if any)
+if [ -f ${RC} ]; then
+    cp ${RC} ${DESTDIR}/usr/etc/rc.local
+    echo "./usr/etc/rc.local type=file uid=0 gid=0 mode=0644" >> ${IMG_DIR}/input
+fi
+
+# add fstab
+echo "./etc/fstab type=file uid=0 gid=0 mode=0755 size=747 time=1365060731.000000000" >> ${IMG_DIR}/input
+
+# fill root.img (skipping /usr entries while keeping the /usr directory)
+cat ${IMG_DIR}/input  | grep -v "^./usr/" | ${CROSS_TOOLS}/nbtoproto -b ${DESTDIR} -o ${IMG_DIR}/root.proto
+
+#
+# Create proto files for /usr and /home using toproto.
+#
+cat ${IMG_DIR}/input  | grep  "^\./usr/\|^. "  | sed "s,\./usr,\.,g" | ${CROSS_TOOLS}/nbtoproto -b ${DESTDIR}/usr -o ${IMG_DIR}/usr.proto
+cat ${IMG_DIR}/input  | grep  "^\./home/\|^. "  | sed "s,\./home,\.,g" | ${CROSS_TOOLS}/nbtoproto -b ${DESTDIR}/home -o ${IMG_DIR}/home.proto
+
+#
+# Create the FAT partition, which contains the bootloader files, kernel and modules
 #
 dd if=/dev/zero of=${IMG_DIR}/fat.img bs=512 count=1 seek=$(($FAT_SIZE -1)) 2>/dev/null
-
-#
-# Create the empty image where we later will put the partitions in.
-#
-rm -f ${IMG}
-dd if=/dev/zero of=${IMG} bs=512 count=1 seek=$(($IMG_SIZE -1))
-
-#
-# Do some math to determine the start addresses of the partitions.
-# Don't leave holes so the 'partition' invocation later is easy.
-#
-FAT_START=2048
-ROOT_START=$(($FAT_START + $FAT_SIZE))
-HOME_START=$(($ROOT_START + $ROOT_SIZE))
-USR_START=$(($HOME_START + $HOME_SIZE))
-
-#
-# Write the partition table using the natively compiled
-# minix partition utility
-#
-${CROSS_TOOLS}/nbpartition -m ${IMG} ${FAT_START} "c:${FAT_SIZE}*" 81:${ROOT_SIZE} 81:${HOME_SIZE} 81:${USR_SIZE}
 
 #
 # Format the fat partition and put the bootloaders
@@ -184,8 +226,8 @@ ${MKFS_VFAT_CMD} ${MKFS_VFAT_OPTS} ${IMG_DIR}/fat.img
 # -p add a prefix to the network booted files (e.g. xm/"
 # -c set console e.g. tty02 or tty00
 # -v set verbosity e.g. 0 to 3
-#./releasetools/gen_uEnv.txt.sh -c ${CONSOLE} -n -p bb/ > ${IMG_DIR}/uEnv.txt
-./releasetools/gen_uEnv.txt.sh -c ${CONSOLE}  > ${IMG_DIR}/uEnv.txt
+#${RELEASETOOLSDIR}/gen_uEnv.txt.sh -c ${CONSOLE} -n -p bb/ > ${IMG_DIR}/uEnv.txt
+${RELEASETOOLSDIR}/gen_uEnv.txt.sh -c ${CONSOLE}  > ${IMG_DIR}/uEnv.txt
 
 echo "Copying configuration kernel and boot modules"
 mcopy -bsp -i ${IMG_DIR}/fat.img  ${IMG_DIR}/$MLO ::MLO
@@ -193,13 +235,8 @@ mcopy -bsp -i ${IMG_DIR}/fat.img ${IMG_DIR}/$UBOOT ::u-boot.img
 mcopy -bsp -i ${IMG_DIR}/fat.img ${IMG_DIR}/uEnv.txt ::uEnv.txt
 
 #
-# For tftp booting
-#
-cp ${IMG_DIR}/uEnv.txt ${OBJ}/
-
-#
-# Do some last processing of the kernel and servers before also putting
-# them on the FAT
+# Do some last processing of the kernel and servers and then put them on the FAT
+# partition.
 #
 ${CROSS_PREFIX}objcopy ${OBJ}/minix/kernel/kernel -O binary ${OBJ}/kernel.bin
 
@@ -224,42 +261,35 @@ do
 done
 
 #
-# make the different file system. this part is *also* hacky. We first convert
-# the METALOG.sanitised using mtree into a input METALOG containing uids and
-# gids.
-# Afther that we do some processing to convert the METALOG into a proto file
-# that can be used by mkfs.mfs
+# For tftp booting
 #
-echo "creating the file systems"
+cp ${IMG_DIR}/uEnv.txt ${OBJ}/
 
 #
-# read METALOG and use mtree to conver the user and group names into uid and gids
-# FIX put "input somwhere clean"
+# Create the empty image where we later will put the partitions in.
+# Make sure it is at least 2GB, otherwise the SD card will not be detected
+# correctly in qemu / HW.
 #
-cat ${DESTDIR}/METALOG.sanitised | ${CROSS_TOOLS}/nbmtree -N ${DESTDIR}/etc -C -K device > ${IMG_DIR}/input
+dd if=/dev/zero of=${IMG} bs=512 count=1 seek=$(($IMG_SIZE -1))
 
-# add fstab
-echo "./etc/fstab type=file uid=0 gid=0 mode=0755 size=747 time=1365060731.000000000" >> ${IMG_DIR}/input
-
-# fill root.img (skipping /usr entries while keeping the /usr directory)
-cat ${IMG_DIR}/input  | grep -v "^./usr/" | ${CROSS_TOOLS}/nbtoproto -b ${DESTDIR} -o ${IMG_DIR}/root.proto
-
-#
-# Create proto files for /usr and /home using toproto.
-#
-cat ${IMG_DIR}/input  | grep  "^\./usr/\|^. "  | sed "s,\./usr,\.,g" | ${CROSS_TOOLS}/nbtoproto -b ${DESTDIR}/usr -o ${IMG_DIR}/usr.proto
-cat ${IMG_DIR}/input  | grep  "^\./home/\|^. "  | sed "s,\./home,\.,g" | ${CROSS_TOOLS}/nbtoproto -b ${DESTDIR}/home -o ${IMG_DIR}/home.proto
 
 #
 # Generate /root, /usr and /home partition images.
 #
 echo "Writing Minix filesystem images"
 echo " - ROOT"
-${CROSS_TOOLS}/nbmkfs.mfs -I $((${ROOT_START} * 512)) -b $((${ROOT_SIZE} / 8)) ${IMG} ${IMG_DIR}/root.proto
+_ROOT_SIZE=$((`${CROSS_TOOLS}/nbmkfs.mfs -I $((${ROOT_START} * 512)) -b $((${ROOT_SIZE} / 8)) ${IMG} ${IMG_DIR}/root.proto`/512))
 echo " - USR"
-${CROSS_TOOLS}/nbmkfs.mfs -I $((${USR_START} * 512)) -b $((${USR_SIZE} / 8))  ${IMG}  ${IMG_DIR}/usr.proto
+_USR_SIZE=$((`${CROSS_TOOLS}/nbmkfs.mfs -I $((${USR_START} * 512)) -b $((${USR_SIZE} / 8))  ${IMG}  ${IMG_DIR}/usr.proto`/512))
 echo " - HOME"
-${CROSS_TOOLS}/nbmkfs.mfs -I $((${HOME_START} * 512)) -b $((${HOME_SIZE} / 8)) ${IMG} ${IMG_DIR}/home.proto
+_HOME_SIZE=$((`${CROSS_TOOLS}/nbmkfs.mfs -I $((${HOME_START} * 512)) -b $((${HOME_SIZE} / 8)) ${IMG} ${IMG_DIR}/home.proto`/512))
+
+#
+# Write the partition table using the natively compiled
+# minix partition utility
+#
+${CROSS_TOOLS}/nbpartition -m ${IMG} ${FAT_START} "c:${FAT_SIZE}*" \
+	81:${_ROOT_SIZE} 81:${_USR_SIZE} 81:${_HOME_SIZE}
 
 #
 # Merge the partitions into a single image.
